renaming cli file
