+++ gnmi_server/platform_system_health_cli_test.go
